requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
requests>=2.31.0
openai>=1.3.0
clarifai
supabase
plotly
streamlit-option-menu
firebase-admin